refactor: move go2proto configuration from flags to comment directives (#3832)
This makes the source files more self-contained, which is nice, but is
also a step towards being able to support multiple packages - when we
import a Go package with directives, we will know that it is a protobuf
package, and we won't generate types for it in the current package.

Next I'll move the definition of the "roots" into code.