fix: use proto to communicate with Hot Reload process (#4076)
The intention here is that gRPC is now used to communicate between the
JVM hot reload process and the plugin. This will allow FTL to have more
control over the process, rather than relying on the Quarkus native hot
reload behaviour.

This is stage one of the refactoring, it removes the runner-info file
and move responsibility for initiating reload to the plugin.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>