chore: use a single sumtype for all schema modifying events (#4178)
Unifies all schema changing operations in provisioner and controller to
go through same events and interface.

The state in provisioner is still separate from controller for now. They
should be easy to put together after this.

Changesets will change how the state is managed quite a bit, but should
be easy to implement here

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>