chore: use grpc client when calling schema service from controller (#4198)
Adds a new endpoint for getting deployments to the schemaservice, and
uses that in controller.
We will probably want to refactor this endpoint soon when we look at
changesets.

This is a minimal change to get the controller to use the grpc client,
any refactoring on the schema service APIs will follow after this.