fix: update provisioner config in charts (#4277)
Adds fallback to runner scaling if runner provisioner is unspecified.
This is a bit of a hack, but currently k8s scaling is not a provisioner
plugin.

Not sure how to test this without committing/releasing/trying.
It looks like "controller" and "kubernetes" are not provisioner plugins
anymore. K8s scaling is always added when running `ftl-provisioner` so
that hopefully handles the k8s part. There are no resources named
`controller` so I think that's old as well.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>