feat: Raft Ops (#4347)
Generates a random replica id for non initial members. Properly restarts
the pod if the id is taken. Makes initial member ids adjustable.

After this, and with the help of the Raft API, we should be able to
recover any corrupted data nodes (as long as we retain quorum)

Next, I will write a quick runbook on managing the Raft cluster.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>