refactor: Refactor Cloudformation Provisioner (#4461)
Separates the CF provisioning into stages with specific inputs and
outputs.

This will allow us to overwrite some outputs from CF execution with
either hard coded values or executions from a shared CF stack to use a
common RDS instance for DBs if desired.

On a high level, the idea here is that we define an ordered list of
`Stage`s to execute, and a set of `Ecxecutor`s applying to each stage,
and a set of resource types, converting a `State` instance in that stage
to a state in a followup `Stage`

So, a Postgres provisioner gets added to the CF template in
`StageProvisioning` and the virtual DB and user are created in
`StageSetup` after which we reach the final stage `StageDone` from which
the runtime updates are created.

This allows us to configure optional executors for `StageProvisioning`
in the future while still use rest of the provisioning tasks unchanged.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>