fix: prevent immediate execution of new migrations (#4545)
Steps to repro:
- Run `ftl dev`
- Create a module
- Define a db in the module
- Run `ftl new-sql-migration <modulename>.<dbname> <migrationname>`

The new migration would be created and then FTL will immediately detect
it and deploy the module, causing the empty migration to be completed.
Writing the actual migration now does nothing...

Hacky fix: We generate the migrartion with placeholder text which makes
the migration invalid. This causes the automatic build to fail,
preventing the migration until the user updates it with an actual
migration.