feat: new config package (#4862)
This basically rewrites and simplifies the existing configuration
package. I was going to do the migration in one PR, but this is already
getting insanely large.

More details:

- Switch profiles to use new config package
- Introduce config.ProviderKey. This supports encoding payloads along
with the kind of provider, allowing configuration of the provider during
initialisation. eg. `op:<vault>` encodes the 1Password vault within the
key.
- Add remote config provider using AdminService. Not fully implemented
as the AdminService will be changed soon.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>