fix: implement a ftl logs command to allow streaming of logs from modules
For now, it streams all logs from the start, and only streaming is supported.
Will add a limit of how much logs to read as a follow-up, as well just returning latest logs.

Also, timeline does scan all events on every update to the event stream. Will look at fixing this as aa follow-up as well