feat: MCP reads and writes files and can proactively provide info (#4893)
Aims:
- Goose takes a lot of time between each action taken. Often we know the
thing goose should find out next, so we should just proactively gather
that info for goose to avoid needless back and forth with goose
- Goose is given hints as to what is in code files via the schema.
Sometimes goose is confident enough to overwrite those files without
actually **reading** those files first... which means it just makes up
the parts of the file it wasn't trying to change. We want some level of
guarantee that goose reads a file before writing the file.

Changes:
- Added Read and Write MCP tools for FTL, which we tell Goose to use
instead of the general `develop` tools included with Goose.
- Read tool returns the content as well as a verification token (just a
hash of the content)
- Write tool requires a verification token to make sure the content has
been read and understood before being altered.
- If a verification token is not provided, or it doesnt match the
content on disk, the Write fails, but we also return a read response (ie
the new content and a new token)
- If the write succeeds we also provide a new token to avoid having to
call the Read tool again (useful when going back and forth fixing
issues).
- MCP tools can also include the result of the FTL Status tool to avoid
additional back and forths with goose.
- eg: writing a file, or creating a module, will wait for the new build
to finish and report to goose the latest schema / build errors.
- MCP tools can also have detection of files that should be read and
provided to goose ahead of time.
- For tools this is enabled for, any output is checked for file paths
within the project.
- FTL will invoke the Read tool for each of these files and provide it
to goose without being asked.
- This speeds up some workflows with goose and also helps goose do the
right thing.
- eg: when creating a new mysql database, FTL will automatically read in
the initial migration file with the token that can be used to write to
that file. Same for creating new migrations.
- We're now better at what info we provide goose vs what we show the
user.
- Write tool: we display a diff to the user to make it obvious what
changed. This is not fed into the LLM so it doesnt waste tokens
- Read tool: we don't print the whole file into the console (it was hard
to follow along with what goose was doing when this happened). Not sure
what we should print here though, currently it just prints the file name
being read.
- All this info we are proactively providing to goose does not show up
to the user.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>