feat(Juho): realm in schema (#5053)
For now, we assume there is always exactly one realm in schema, the
internal realm of the cluster.
However, this creates a foundation for supporting realm federation and
declaring external egress realms to the schema.

This changes `ftl get schema` to return 
```
realm default { 
  builtin module builtin {
    ...
  }

  module time {
    ...
  }

  module echo {
     ...
  }
}
```

The underlying state or changesets have not changed yet, that is a
follow-up.

After the changesets and the state support realms, we can change the
"default" name to the name from `ftl-project.toml`

modules are still referred my module name in many places, as a follow-up
we need to move to a structured moduleRef that takes the realm into
account.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>