chore: switch to github.com/alecthomas/errors to get stack traces
`fmt.Errorf()` is idiomatic, but in practice sucks because we don't get stack
traces. This PR switches us to alecthomas/errors which was pulled out of
cashapp/hermit a couple of years ago.

The major difference between it and `fmt.Errorf()` is that only `fmt.Errorf()`
can support `%w`, so to wrap errors we have to now use `errors.Wrap{,f}(err, ...)`.

Use `fmt.Printf("%+v", err)` to print stack traces of an error, or `export DEBUG=1`.