fix: connection errors do not count as a consumption attempt (#5018)
fixes https://github.com/block/ftl/issues/4822
Confirmed that killing FTL could cause progress consumer group offsets.
It is a very hard edge case to reproduce though.
Steps:
- Call to verb fails due to a `connect.Error` while the pubsub context
is not yet cancelled.
- Consumer marks the offset
- Sarama library then tries to sync the offset to kafka/redpanda. This
usually never actually completes which is why we don't see this happen
in practice a lot.

Was not able to make an integration test that would fail for this.