chore: make previously flaky pubsub test strict again (#5172)
closes https://github.com/block/ftl/issues/4831
`TestConsumerGroupMembership` tries to make sure that runners
de-register themselves from redpanda on shutdown.
That test was flakey though, and so the duration was made more leniant.
But that meant the test didn't really check anything as the new timeout
was longer than redpandas timeout, so if our runners didn't notify
redpanda the test would still pass.

A recent fix to how we handle shutdowns while consuming may have fixed
the flakiness, so we can make the test strict again.
Please revert this PR if the test remains flakey...