feat: mcp returns diffs of generated sql interfaces (#5198)
closes https://github.com/block/ftl/issues/5178
Previously we had a hard-coded way for go modules where whenever a db
query or schema was updated by the MCP, we would return the latest
language-specific interface (`queries.ftl.go`) back as well so that the
LLM would know how to write code against it.

Changes:
- Allows multiple languages to support it by adding a language plugin
command called `GetSQLInterfaces` which returns a mapping from
declaration name to the declaration
- The MCP now diffs the SQL interfaces from the language plugin and only
notifies the LLM of added/updated/removed declarations. This will
greatly reduce token usage when there are lots of queries defined.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>