fix: pass log messages through directly in the runner (#5425)
At the moment we don't have a proper non-dev mode translation layer
between the JVM and go log JSON formats, so the JVM logs are getting all
of their exception information being stripped from them. This can hide
critical errors, as only the outermost exception is present in the logs.

For now the simplest solution is to just pass them through so they get
written directly to the kube pod logs. There is already transaction in
place in dev mode so this only affects prod.